...
git-svn-id: svn+ssh://src.fovea.cc/svn/flobopuyo/branches/autotools@1236 053fdb51-1f17-0410-be66-d00040120cd6

+++ goomsl/goomsl_heap.c
@@ -1,5 +1,6 @@
 #include "goomsl_heap.h"
 #include <stdlib.h>
+#include <stdint.h>
 
 struct _GOOM_HEAP {
   void **arrays;
@@ -39,8 +40,8 @@ static void align_it(GoomHeap *_this, int alignment)
 {
   if ((alignment > 1) && (_this->number_of_arrays>0)) {
     void *last_array = _this->arrays[_this->number_of_arrays - 1];
-    int   last_address = (int)last_array + _this->consumed_in_last_array;
-    int   decal = (last_address % alignment);
+    intptr_t last_address = (intptr_t)last_array + _this->consumed_in_last_array;
+    intptr_t decal = (last_address % alignment);
     if (decal != 0) {
       _this->consumed_in_last_array += alignment - decal;
     }
